create Collatz_Conjecture/Dependency.py
